refactor: extract file download
